Clean up str/bytes iterator code
